# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Insulin Dose Calculator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Profile {
  id: number;
  /** Grams of carbs per 1 unit of insulin */
  carbRatio: number;
  /** How much 1 unit of insulin lowers blood glucose (mg/dL) */
  correctionFactor: number;
  /** Target blood glucose level (mg/dL) */
  targetBg: number;
  updatedAt: string;
}

export interface UpdateProfileBody {
  /** Grams of carbs per 1 unit of insulin */
  carbRatio: number;
  /** How much 1 unit of insulin lowers blood glucose (mg/dL) */
  correctionFactor: number;
  /** Target blood glucose level (mg/dL) */
  targetBg: number;
}

export interface CalculateDoseBody {
  /** Grams of carbohydrates in the meal */
  carbs: number;
  /** Current blood glucose level (mg/dL) */
  currentBg: number;
}

export interface DoseResult {
  /** Insulin units for carbs */
  carbDose: number;
  /** Insulin units for blood glucose correction */
  correctionDose: number;
  /** Total insulin units to administer (never negative, capped at 30) */
  totalDose: number;
  /** Whether the dose was capped for safety */
  isCapped: boolean;
  carbs: number;
  currentBg: number;
  carbRatio: number;
  correctionFactor: number;
  targetBg: number;
}

export interface CalculationRecord {
  id: number;
  carbs: number;
  currentBg: number;
  carbDose: number;
  correctionDose: number;
  totalDose: number;
  carbRatio: number;
  correctionFactor: number;
  targetBg: number;
  isCapped: boolean;
  createdAt: string;
}

export interface CalculationSummary {
  totalCalculations: number;
  /** @nullable */
  averageTotalDose: number | null;
  /** @nullable */
  averageCarbs: number | null;
  /** @nullable */
  averageCurrentBg: number | null;
  /** Number of calculations in the last 7 days */
  recentCalculationsCount: number;
}

export type ListCalculationsParams = {
  limit?: number;
};