# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Insulin Dose Calculator API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: profile
    description: User profile and insulin settings
  - name: calculations
    description: Insulin dose calculations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      description: Returns the user's insulin settings profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      description: Update insulin settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /calculate:
    post:
      operationId: calculateDose
      tags: [calculations]
      summary: Calculate insulin dose
      description: Calculate the insulin dose based on carbs and current blood glucose
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalculateDoseBody"
      responses:
        "200":
          description: Calculated dose
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoseResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /calculations:
    get:
      operationId: listCalculations
      tags: [calculations]
      summary: List past calculations
      description: Returns history of past insulin dose calculations
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of calculations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalculationRecord"

  /calculations/summary:
    get:
      operationId: getCalculationSummary
      tags: [calculations]
      summary: Get calculation summary stats
      description: Returns aggregated stats about past calculations
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculationSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Profile:
      type: object
      properties:
        id:
          type: integer
        carbRatio:
          type: number
          description: Grams of carbs per 1 unit of insulin
        correctionFactor:
          type: number
          description: How much 1 unit of insulin lowers blood glucose (mg/dL)
        targetBg:
          type: number
          description: Target blood glucose level (mg/dL)
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - carbRatio
        - correctionFactor
        - targetBg
        - updatedAt

    UpdateProfileBody:
      type: object
      properties:
        carbRatio:
          type: number
          description: Grams of carbs per 1 unit of insulin
        correctionFactor:
          type: number
          description: How much 1 unit of insulin lowers blood glucose (mg/dL)
        targetBg:
          type: number
          description: Target blood glucose level (mg/dL)
      required:
        - carbRatio
        - correctionFactor
        - targetBg

    CalculateDoseBody:
      type: object
      properties:
        carbs:
          type: number
          description: Grams of carbohydrates in the meal
        currentBg:
          type: number
          description: Current blood glucose level (mg/dL)
      required:
        - carbs
        - currentBg

    DoseResult:
      type: object
      properties:
        carbDose:
          type: number
          description: Insulin units for carbs
        correctionDose:
          type: number
          description: Insulin units for blood glucose correction
        totalDose:
          type: number
          description: Total insulin units to administer (never negative, capped at 30)
        isCapped:
          type: boolean
          description: Whether the dose was capped for safety
        carbs:
          type: number
        currentBg:
          type: number
        carbRatio:
          type: number
        correctionFactor:
          type: number
        targetBg:
          type: number
      required:
        - carbDose
        - correctionDose
        - totalDose
        - isCapped
        - carbs
        - currentBg
        - carbRatio
        - correctionFactor
        - targetBg

    CalculationRecord:
      type: object
      properties:
        id:
          type: integer
        carbs:
          type: number
        currentBg:
          type: number
        carbDose:
          type: number
        correctionDose:
          type: number
        totalDose:
          type: number
        carbRatio:
          type: number
        correctionFactor:
          type: number
        targetBg:
          type: number
        isCapped:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - carbs
        - currentBg
        - carbDose
        - correctionDose
        - totalDose
        - carbRatio
        - correctionFactor
        - targetBg
        - isCapped
        - createdAt

    CalculationSummary:
      type: object
      properties:
        totalCalculations:
          type: integer
        averageTotalDose:
          type: ["number", "null"]
        averageCarbs:
          type: ["number", "null"]
        averageCurrentBg:
          type: ["number", "null"]
        recentCalculationsCount:
          type: integer
          description: Number of calculations in the last 7 days
      required:
        - totalCalculations
        - averageTotalDose
        - averageCarbs
        - averageCurrentBg
        - recentCalculationsCount