# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Insulin Dose Calculator API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns the user's insulin settings profile
 * @summary Get user profile
 */
export const GetProfileResponse = zod.object({
  id: zod.number(),
  carbRatio: zod.number().describe("Grams of carbs per 1 unit of insulin"),
  correctionFactor: zod
    .number()
    .describe("How much 1 unit of insulin lowers blood glucose (mg\/dL)"),
  targetBg: zod.number().describe("Target blood glucose level (mg\/dL)"),
  updatedAt: zod.coerce.date(),
});

/**
 * Update insulin settings
 * @summary Update user profile
 */
export const UpdateProfileBody = zod.object({
  carbRatio: zod.number().describe("Grams of carbs per 1 unit of insulin"),
  correctionFactor: zod
    .number()
    .describe("How much 1 unit of insulin lowers blood glucose (mg\/dL)"),
  targetBg: zod.number().describe("Target blood glucose level (mg\/dL)"),
});

export const UpdateProfileResponse = zod.object({
  id: zod.number(),
  carbRatio: zod.number().describe("Grams of carbs per 1 unit of insulin"),
  correctionFactor: zod
    .number()
    .describe("How much 1 unit of insulin lowers blood glucose (mg\/dL)"),
  targetBg: zod.number().describe("Target blood glucose level (mg\/dL)"),
  updatedAt: zod.coerce.date(),
});

/**
 * Calculate the insulin dose based on carbs and current blood glucose
 * @summary Calculate insulin dose
 */
export const CalculateDoseBody = zod.object({
  carbs: zod.number().describe("Grams of carbohydrates in the meal"),
  currentBg: zod.number().describe("Current blood glucose level (mg\/dL)"),
});

export const CalculateDoseResponse = zod.object({
  carbDose: zod.number().describe("Insulin units for carbs"),
  correctionDose: zod
    .number()
    .describe("Insulin units for blood glucose correction"),
  totalDose: zod
    .number()
    .describe(
      "Total insulin units to administer (never negative, capped at 30)",
    ),
  isCapped: zod.boolean().describe("Whether the dose was capped for safety"),
  carbs: zod.number(),
  currentBg: zod.number(),
  carbRatio: zod.number(),
  correctionFactor: zod.number(),
  targetBg: zod.number(),
});

/**
 * Returns history of past insulin dose calculations
 * @summary List past calculations
 */
export const listCalculationsQueryLimitDefault = 20;

export const ListCalculationsQueryParams = zod.object({
  limit: zod.coerce.number().default(listCalculationsQueryLimitDefault),
});

export const ListCalculationsResponseItem = zod.object({
  id: zod.number(),
  carbs: zod.number(),
  currentBg: zod.number(),
  carbDose: zod.number(),
  correctionDose: zod.number(),
  totalDose: zod.number(),
  carbRatio: zod.number(),
  correctionFactor: zod.number(),
  targetBg: zod.number(),
  isCapped: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListCalculationsResponse = zod.array(ListCalculationsResponseItem);

/**
 * Returns aggregated stats about past calculations
 * @summary Get calculation summary stats
 */
export const GetCalculationSummaryResponse = zod.object({
  totalCalculations: zod.number(),
  averageTotalDose: zod.number().nullable(),
  averageCarbs: zod.number().nullable(),
  averageCurrentBg: zod.number().nullable(),
  recentCalculationsCount: zod
    .number()
    .describe("Number of calculations in the last 7 days"),
});