# lib/api-zod/src/generated/types/calculateDoseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Insulin Dose Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface CalculateDoseBody {
  /** Grams of carbohydrates in the meal */
  carbs: number;
  /** Current blood glucose level (mg/dL) */
  currentBg: number;
}