# lib/api-zod/src/generated/types/calculationRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Insulin Dose Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface CalculationRecord {
  id: number;
  carbs: number;
  currentBg: number;
  carbDose: number;
  correctionDose: number;
  totalDose: number;
  carbRatio: number;
  correctionFactor: number;
  targetBg: number;
  isCapped: boolean;
  createdAt: Date;
}