# lib/api-zod/src/generated/types/calculationSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Insulin Dose Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface CalculationSummary {
  totalCalculations: number;
  /** @nullable */
  averageTotalDose: number | null;
  /** @nullable */
  averageCarbs: number | null;
  /** @nullable */
  averageCurrentBg: number | null;
  /** Number of calculations in the last 7 days */
  recentCalculationsCount: number;
}