# lib/api-zod/src/generated/types/doseResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Insulin Dose Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface DoseResult {
  /** Insulin units for carbs */
  carbDose: number;
  /** Insulin units for blood glucose correction */
  correctionDose: number;
  /** Total insulin units to administer (never negative, capped at 30) */
  totalDose: number;
  /** Whether the dose was capped for safety */
  isCapped: boolean;
  carbs: number;
  currentBg: number;
  carbRatio: number;
  correctionFactor: number;
  targetBg: number;
}