# lib/api-zod/src/generated/types/updateProfileBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Insulin Dose Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProfileBody {
  /** Grams of carbs per 1 unit of insulin */
  carbRatio: number;
  /** How much 1 unit of insulin lowers blood glucose (mg/dL) */
  correctionFactor: number;
  /** Target blood glucose level (mg/dL) */
  targetBg: number;
}